fix: define lp types only in cf-rpc-types
