+++ api/bin/chainflip-lp-api/src/main.rs
@@ -1,4 +1,3 @@
-use crate::rpc_types::CloseOrderJson;
 use anyhow::anyhow;
 use cf_primitives::{BasisPoints, BlockNumber, EgressId};
 use cf_utilities::{
@@ -10,8 +9,8 @@ use cf_utilities::{
 use chainflip_api::{
 	self,
 	lp::{
-		types::{LimitOrRangeOrder, LimitOrder, RangeOrder},
-		ApiWaitForResult, LpApi, Side, Tick,
+		ApiWaitForResult, CloseOrderJson, LimitOrRangeOrder, LimitOrder, LpApi, OpenSwapChannels,
+		OrderIdJson, RangeOrder, RangeOrderSizeJson, Side, Tick,
 	},
 	primitives::{
 		chains::{assets::any::AssetMap, Bitcoin, Ethereum, Polkadot},
@@ -31,8 +30,7 @@ use jsonrpsee::{
 	types::{ErrorCode, ErrorObject, ErrorObjectOwned},
 	PendingSubscriptionSink,
 };
-use pallet_cf_pools::{CloseOrder, IncreaseOrDecrease, OrderId, RangeOrderSize, MAX_ORDERS_DELETE};
-use rpc_types::{OpenSwapChannels, OrderIdJson, RangeOrderSizeJson};
+use pallet_cf_pools::{CloseOrder, IncreaseOrDecrease, MAX_ORDERS_DELETE};
 use sp_core::{bounded::BoundedVec, ConstU32, H256, U256};
 use std::{
 	ops::Range,
@@ -40,81 +38,6 @@ use std::{
 	sync::{atomic::AtomicBool, Arc},
 };
 use tracing::log;
-
-/// Contains RPC interface types that differ from internal types.
-pub mod rpc_types {
-	use super::*;
-	use anyhow::anyhow;
-	use cf_primitives::chains::assets::any;
-	use cf_utilities::rpc::NumberOrHex;
-	use chainflip_api::{lp::PoolPairsMap, queries::SwapChannelInfo};
-	use serde::{Deserialize, Serialize};
-
-	#[derive(Copy, Clone, Debug, Serialize, Deserialize)]
-	pub struct OrderIdJson(NumberOrHex);
-	impl TryFrom<OrderIdJson> for OrderId {
-		type Error = anyhow::Error;
-
-		fn try_from(value: OrderIdJson) -> Result<Self, Self::Error> {
-			value.0.try_into().map_err(|_| anyhow!("Failed to convert order id to u64"))
-		}
-	}
-
-	#[derive(Copy, Clone, Debug, Serialize, Deserialize)]
-	pub enum RangeOrderSizeJson {
-		AssetAmounts { maximum: PoolPairsMap<NumberOrHex>, minimum: PoolPairsMap<NumberOrHex> },
-		Liquidity { liquidity: NumberOrHex },
-	}
-	impl TryFrom<RangeOrderSizeJson> for RangeOrderSize {
-		type Error = anyhow::Error;
-
-		fn try_from(value: RangeOrderSizeJson) -> Result<Self, Self::Error> {
-			Ok(match value {
-				RangeOrderSizeJson::AssetAmounts { maximum, minimum } =>
-					RangeOrderSize::AssetAmounts {
-						maximum: maximum
-							.try_map(TryInto::try_into)
-							.map_err(|_| anyhow!("Failed to convert maximums to u128"))?,
-						minimum: minimum
-							.try_map(TryInto::try_into)
-							.map_err(|_| anyhow!("Failed to convert minimums to u128"))?,
-					},
-				RangeOrderSizeJson::Liquidity { liquidity } => RangeOrderSize::Liquidity {
-					liquidity: liquidity
-						.try_into()
-						.map_err(|_| anyhow!("Failed to convert liquidity to u128"))?,
-				},
-			})
-		}
-	}
-
-	#[derive(Serialize, Deserialize, Clone)]
-	pub struct OpenSwapChannels {
-		pub ethereum: Vec<SwapChannelInfo<Ethereum>>,
-		pub bitcoin: Vec<SwapChannelInfo<Bitcoin>>,
-		pub polkadot: Vec<SwapChannelInfo<Polkadot>>,
-	}
-
-	#[derive(Copy, Clone, Debug, Serialize, Deserialize)]
-	#[serde(untagged)]
-	pub enum CloseOrderJson {
-		Limit { base_asset: any::Asset, quote_asset: any::Asset, side: Side, id: OrderIdJson },
-		Range { base_asset: any::Asset, quote_asset: any::Asset, id: OrderIdJson },
-	}
-
-	impl TryFrom<CloseOrderJson> for CloseOrder {
-		type Error = anyhow::Error;
-
-		fn try_from(value: CloseOrderJson) -> Result<Self, Self::Error> {
-			Ok(match value {
-				CloseOrderJson::Limit { base_asset, quote_asset, side, id } =>
-					CloseOrder::Limit { base_asset, quote_asset, side, id: id.try_into()? },
-				CloseOrderJson::Range { base_asset, quote_asset, id } =>
-					CloseOrder::Range { base_asset, quote_asset, id: id.try_into()? },
-			})
-		}
-	}
-}
 
 #[rpc(server, client, namespace = "lp")]
 pub trait Rpc {
