+++ api/lib/src/lp.rs
@@ -18,83 +18,21 @@ use chainflip_engine::state_chain_observer::client::{
 };
 use frame_support::{pallet_prelude::ConstU32, BoundedVec};
 use pallet_cf_pools::{CloseOrder, IncreaseOrDecrease, OrderId, RangeOrderSize, MAX_ORDERS_DELETE};
-use serde::{Deserialize, Serialize};
-use sp_core::{H256, U256};
+use sp_core::H256;
 use state_chain_runtime::RuntimeCall;
 use std::ops::Range;
-use types::LimitOrRangeOrder;
 
-#[derive(Debug, Serialize, Deserialize, Clone)]
-#[serde(rename_all = "snake_case")]
-pub enum ApiWaitForResult<T> {
-	TxHash(H256),
-	TxDetails { tx_hash: H256, response: T },
-}
-
-impl<T> ApiWaitForResult<T> {
-	pub fn map_details<R>(self, f: impl FnOnce(T) -> R) -> ApiWaitForResult<R> {
-		match self {
-			ApiWaitForResult::TxHash(hash) => ApiWaitForResult::TxHash(hash),
-			ApiWaitForResult::TxDetails { response, tx_hash } =>
-				ApiWaitForResult::TxDetails { tx_hash, response: f(response) },
-		}
-	}
-
-	#[track_caller]
-	pub fn unwrap_details(self) -> T {
-		match self {
-			ApiWaitForResult::TxHash(_) => panic!("unwrap_details called on TransactionHash"),
-			ApiWaitForResult::TxDetails { response, .. } => response,
-		}
-	}
-}
-
-pub mod types {
-	use super::*;
-
-	#[derive(Serialize, Deserialize, Clone)]
-	pub struct RangeOrder {
-		pub base_asset: Asset,
-		pub quote_asset: Asset,
-		pub id: U256,
-		pub tick_range: Range<Tick>,
-		pub liquidity_total: U256,
-		pub collected_fees: PoolPairsMap<U256>,
-		pub size_change: Option<IncreaseOrDecrease<RangeOrderChange>>,
-	}
-
-	#[derive(Serialize, Deserialize, Clone)]
-	pub struct RangeOrderChange {
-		pub liquidity: U256,
-		pub amounts: PoolPairsMap<U256>,
-	}
-
-	#[derive(Serialize, Deserialize, Clone)]
-	pub struct LimitOrder {
-		pub base_asset: Asset,
-		pub quote_asset: Asset,
-		pub side: Side,
-		pub id: U256,
-		pub tick: Tick,
-		pub sell_amount_total: U256,
-		pub collected_fees: U256,
-		pub bought_amount: U256,
-		pub sell_amount_change: Option<IncreaseOrDecrease<U256>>,
-	}
-
-	#[derive(Serialize, Deserialize, Clone)]
-	pub enum LimitOrRangeOrder {
-		LimitOrder(LimitOrder),
-		RangeOrder(RangeOrder),
-	}
-}
+pub use cf_rpc_types::lp::{
+	ApiWaitForResult, CloseOrderJson, LimitOrRangeOrder, LimitOrder, OpenSwapChannels, OrderIdJson,
+	RangeOrder, RangeOrderChange, RangeOrderSizeJson,
+};
 
 fn collect_range_order_returns(
 	events: impl IntoIterator<Item = state_chain_runtime::RuntimeEvent>,
-) -> Vec<types::RangeOrder> {
+) -> Vec<RangeOrder> {
 	filter_orders(events)
 		.filter_map(|order| match order {
-			types::LimitOrRangeOrder::RangeOrder(range_order) => Some(range_order),
+			LimitOrRangeOrder::RangeOrder(range_order) => Some(range_order),
 			_ => None,
 		})
 		.collect()
@@ -102,10 +40,10 @@ fn collect_range_order_returns(
 
 fn collect_limit_order_returns(
 	events: impl IntoIterator<Item = state_chain_runtime::RuntimeEvent>,
-) -> Vec<types::LimitOrder> {
+) -> Vec<LimitOrder> {
 	filter_orders(events)
 		.filter_map(|order| match order {
-			types::LimitOrRangeOrder::LimitOrder(limit_order) => Some(limit_order),
+			LimitOrRangeOrder::LimitOrder(limit_order) => Some(limit_order),
 			_ => None,
 		})
 		.collect()
@@ -113,7 +51,7 @@ fn collect_limit_order_returns(
 
 fn collect_order_returns(
 	events: impl IntoIterator<Item = state_chain_runtime::RuntimeEvent>,
-) -> Vec<types::LimitOrRangeOrder> {
+) -> Vec<LimitOrRangeOrder> {
 	filter_orders(events).collect()
 }
 
@@ -134,7 +72,7 @@ fn filter_orders(
 				id,
 				..
 			},
-		) => Some(types::LimitOrRangeOrder::LimitOrder(types::LimitOrder {
+		) => Some(LimitOrRangeOrder::LimitOrder(LimitOrder {
 			base_asset,
 			quote_asset,
 			side,
@@ -157,12 +95,12 @@ fn filter_orders(
 				id,
 				..
 			},
-		) => Some(types::LimitOrRangeOrder::RangeOrder(types::RangeOrder {
+		) => Some(LimitOrRangeOrder::RangeOrder(RangeOrder {
 			base_asset,
 			quote_asset,
 			id: id.into(),
 			size_change: size_change.map(|increase_or_decrease| {
-				increase_or_decrease.map(|range_order_change| types::RangeOrderChange {
+				increase_or_decrease.map(|range_order_change| RangeOrderChange {
 					liquidity: range_order_change.liquidity.into(),
 					amounts: range_order_change.amounts.map(|amount| amount.into()),
 				})
@@ -325,7 +263,7 @@ pub trait LpApi: SignedExtrinsicApi + Sized + Send + Sync + 'static {
 		option_tick_range: Option<Range<Tick>>,
 		size_change: IncreaseOrDecrease<RangeOrderSize>,
 		wait_for: WaitFor,
-	) -> Result<ApiWaitForResult<Vec<types::RangeOrder>>> {
+	) -> Result<ApiWaitForResult<Vec<RangeOrder>>> {
 		// Submit the mint order
 		Ok(into_api_wait_for_result(
 			self.submit_signed_extrinsic_wait_for(
@@ -351,7 +289,7 @@ pub trait LpApi: SignedExtrinsicApi + Sized + Send + Sync + 'static {
 		option_tick_range: Option<Range<Tick>>,
 		size: RangeOrderSize,
 		wait_for: WaitFor,
-	) -> Result<ApiWaitForResult<Vec<types::RangeOrder>>> {
+	) -> Result<ApiWaitForResult<Vec<RangeOrder>>> {
 		// Submit the mint order
 		Ok(into_api_wait_for_result(
 			self.submit_signed_extrinsic_wait_for(
@@ -379,7 +317,7 @@ pub trait LpApi: SignedExtrinsicApi + Sized + Send + Sync + 'static {
 		amount_change: IncreaseOrDecrease<AssetAmount>,
 		dispatch_at: Option<BlockNumber>,
 		wait_for: WaitFor,
-	) -> Result<ApiWaitForResult<Vec<types::LimitOrder>>> {
+	) -> Result<ApiWaitForResult<Vec<LimitOrder>>> {
 		self.scheduled_or_immediate(
 			pallet_cf_pools::Call::update_limit_order {
 				base_asset,
@@ -405,7 +343,7 @@ pub trait LpApi: SignedExtrinsicApi + Sized + Send + Sync + 'static {
 		sell_amount: AssetAmount,
 		dispatch_at: Option<BlockNumber>,
 		wait_for: WaitFor,
-	) -> Result<ApiWaitForResult<Vec<types::LimitOrder>>> {
+	) -> Result<ApiWaitForResult<Vec<LimitOrder>>> {
 		self.scheduled_or_immediate(
 			pallet_cf_pools::Call::set_limit_order {
 				base_asset,
@@ -426,7 +364,7 @@ pub trait LpApi: SignedExtrinsicApi + Sized + Send + Sync + 'static {
 		call: pallet_cf_pools::Call<state_chain_runtime::Runtime>,
 		dispatch_at: Option<BlockNumber>,
 		wait_for: WaitFor,
-	) -> Result<ApiWaitForResult<Vec<types::LimitOrder>>> {
+	) -> Result<ApiWaitForResult<Vec<LimitOrder>>> {
 		Ok(into_api_wait_for_result(
 			if let Some(dispatch_at) = dispatch_at {
 				self.submit_signed_extrinsic_wait_for(
@@ -458,7 +396,7 @@ pub trait LpApi: SignedExtrinsicApi + Sized + Send + Sync + 'static {
 		&self,
 		orders: BoundedVec<CloseOrder, ConstU32<MAX_ORDERS_DELETE>>,
 		wait_for: WaitFor,
-	) -> Result<ApiWaitForResult<Vec<types::LimitOrRangeOrder>>> {
+	) -> Result<ApiWaitForResult<Vec<LimitOrRangeOrder>>> {
 		Ok(into_api_wait_for_result(
 			self.submit_signed_extrinsic_wait_for(
 				pallet_cf_pools::Call::cancel_orders_batch { orders },
